feat(issues): Remove unused ignore button
Button is unused, merge dropdown options into the archive button.
