--- static/app/components/actions/ignore.tsx
@@ -1,294 +1,0 @@
-import styled from '@emotion/styled';
-
-import {openModal} from 'sentry/actionCreators/modal';
-import {openConfirmModal} from 'sentry/components/confirm';
-import {Button} from 'sentry/components/core/button';
-import {ButtonBar} from 'sentry/components/core/button/buttonBar';
-import CustomIgnoreCountModal from 'sentry/components/customIgnoreCountModal';
-import CustomIgnoreDurationModal from 'sentry/components/customIgnoreDurationModal';
-import type {MenuItemProps} from 'sentry/components/dropdownMenu';
-import {DropdownMenu} from 'sentry/components/dropdownMenu';
-import {Tooltip} from 'sentry/components/tooltip';
-import {IconChevron} from 'sentry/icons';
-import {t, tn} from 'sentry/locale';
-import type {SelectValue} from 'sentry/types/core';
-import type {GroupStatusResolution, IgnoredStatusDetails} from 'sentry/types/group';
-import {GroupStatus, GroupSubstatus} from 'sentry/types/group';
-import getDuration from 'sentry/utils/duration/getDuration';
-
-const ONE_HOUR = 60;
-
-/**
- * Ignore durations are in munutes
- */
-const IGNORE_DURATIONS = [
-  ONE_HOUR / 2,
-  ONE_HOUR * 2,
-  ONE_HOUR * 6,
-  ONE_HOUR * 24,
-  ONE_HOUR * 24 * 7,
-];
-
-const IGNORE_COUNTS = [1, 10, 100, 1000, 10000, 100000];
-
-const IGNORE_WINDOWS: Array<SelectValue<number>> = [
-  {value: ONE_HOUR, label: t('per hour')},
-  {value: ONE_HOUR * 24, label: t('per day')},
-  {value: ONE_HOUR * 24 * 7, label: t('per week')},
-];
-
-/**
- * Create the dropdown submenus
- */
-export function getIgnoreActions({
-  confirmLabel,
-  confirmMessage,
-  shouldConfirm,
-  onUpdate,
-}: Pick<
-  IgnoreActionProps,
-  'shouldConfirm' | 'confirmMessage' | 'confirmLabel' | 'onUpdate'
->) {
-  const onIgnore = (
-    statusDetails: IgnoredStatusDetails | undefined = {},
-    {bypassConfirm} = {bypassConfirm: false}
-  ) => {
-    openConfirmModal({
-      bypass: bypassConfirm || !shouldConfirm,
-      onConfirm: () =>
-        onUpdate({
-          status: GroupStatus.IGNORED,
-          statusDetails,
-          substatus: GroupSubstatus.ARCHIVED_UNTIL_CONDITION_MET,
-        }),
-      message: confirmMessage?.() ?? null,
-      confirmText: confirmLabel,
-    });
-  };
-
-  const onCustomIgnore = (statusDetails: IgnoredStatusDetails) => {
-    onIgnore(statusDetails, {bypassConfirm: true});
-  };
-
-  const openCustomIgnoreDuration = () =>
-    openModal(deps => (
-      <CustomIgnoreDurationModal
-        {...deps}
-        onSelected={details => onCustomIgnore(details)}
-      />
-    ));
-
-  const openCustomIgnoreCount = () =>
-    openModal(deps => (
-      <CustomIgnoreCountModal
-        {...deps}
-        onSelected={details => onCustomIgnore(details)}
-        label={t('Archive this issue until it occurs again\u2026')}
-        countLabel={t('Number of times')}
-        countName="ignoreCount"
-        windowName="ignoreWindow"
-        windowOptions={IGNORE_WINDOWS}
-      />
-    ));
-
-  const openCustomIgnoreUserCount = () =>
-    openModal(deps => (
-      <CustomIgnoreCountModal
-        {...deps}
-        onSelected={details => onCustomIgnore(details)}
-        label={t('Archive this issue until it affects an additional\u2026')}
-        countLabel={t('Number of users')}
-        countName="ignoreUserCount"
-        windowName="ignoreUserWindow"
-        windowOptions={IGNORE_WINDOWS}
-      />
-    ));
-
-  // Move submenu placement when ignore used in top right menu
-  const dropdownItems: MenuItemProps[] = [
-    {
-      key: 'for',
-      label: t('For\u2026'),
-      isSubmenu: true,
-      children: [
-        ...IGNORE_DURATIONS.map(duration => ({
-          key: `for-${duration}`,
-          label: getDuration(duration * 60),
-          onAction: () => onIgnore({ignoreDuration: duration}),
-        })),
-        {
-          key: 'for-custom',
-          label: t('Custom'),
-          onAction: () => openCustomIgnoreDuration(),
-        },
-      ],
-    },
-    {
-      key: 'until-reoccur',
-      label: t('Until this occurs again\u2026'),
-      isSubmenu: true,
-      children: [
-        ...IGNORE_COUNTS.map(count => ({
-          key: `until-reoccur-${count}-times`,
-          label:
-            count === 1
-              ? t('one time\u2026') // This is intentional as unbalanced string formatters are problematic
-              : tn('%s time\u2026', '%s times\u2026', count),
-          isSubmenu: true,
-          children: [
-            {
-              key: `until-reoccur-${count}-times-from-now`,
-              label: t('from now'),
-              onAction: () => onIgnore({ignoreCount: count}),
-            },
-            ...IGNORE_WINDOWS.map(({value, label}) => ({
-              key: `until-reoccur-${count}-times-from-${label}`,
-              label,
-              onAction: () =>
-                onIgnore({
-                  ignoreCount: count,
-                  ignoreWindow: value,
-                }),
-            })),
-          ],
-        })),
-        {
-          key: 'until-reoccur-custom',
-          label: t('Custom'),
-          onAction: () => openCustomIgnoreCount(),
-        },
-      ],
-    },
-    {
-      key: 'until-affect',
-      label: t('Until this affects an additional\u2026'),
-      isSubmenu: true,
-      children: [
-        ...IGNORE_COUNTS.map(count => ({
-          key: `until-affect-${count}-users`,
-          label:
-            count === 1
-              ? t('one user\u2026') // This is intentional as unbalanced string formatters are problematic
-              : tn('%s user\u2026', '%s users\u2026', count),
-          isSubmenu: true,
-          children: [
-            {
-              key: `until-affect-${count}-users-from-now`,
-              label: t('from now'),
-              onAction: () => onIgnore({ignoreUserCount: count}),
-            },
-            ...IGNORE_WINDOWS.map(({value, label}) => ({
-              key: `until-affect-${count}-users-from-${label}`,
-              label,
-              onAction: () =>
-                onIgnore({
-                  ignoreUserCount: count,
-                  ignoreUserWindow: value,
-                }),
-            })),
-          ],
-        })),
-        {
-          key: 'until-affect-custom',
-          label: t('Custom'),
-          onAction: () => openCustomIgnoreUserCount(),
-        },
-      ],
-    },
-  ];
-  return {dropdownItems, onIgnore};
-}
-
-type IgnoreActionProps = {
-  onUpdate: (params: GroupStatusResolution) => void;
-  className?: string;
-  confirmLabel?: string;
-  confirmMessage?: () => React.ReactNode;
-  disabled?: boolean;
-  isIgnored?: boolean;
-  shouldConfirm?: boolean;
-  size?: 'xs' | 'sm';
-};
-
-function IgnoreActions({
-  onUpdate,
-  disabled,
-  shouldConfirm,
-  confirmMessage,
-  className,
-  size = 'xs',
-  confirmLabel = t('Ignore'),
-  isIgnored = false,
-}: IgnoreActionProps) {
-  if (isIgnored) {
-    return (
-      <Tooltip title={t('Change status to unresolved')}>
-        <Button
-          priority="primary"
-          size="xs"
-          onClick={() =>
-            onUpdate({
-              status: GroupStatus.UNRESOLVED,
-              statusDetails: {},
-              substatus: GroupSubstatus.ONGOING,
-            })
-          }
-          aria-label={t('Unignore')}
-        />
-      </Tooltip>
-    );
-  }
-
-  const {dropdownItems, onIgnore} = getIgnoreActions({
-    confirmLabel,
-    onUpdate,
-    shouldConfirm,
-    confirmMessage,
-  });
-
-  return (
-    <ButtonBar merged>
-      <IgnoreButton
-        size={size}
-        className={className}
-        tooltipProps={{delay: 300, disabled}}
-        title={t(
-          'Silences alerts for this issue and removes it from the issue stream by default.'
-        )}
-        onClick={() => onIgnore()}
-        disabled={disabled}
-      >
-        {t('Ignore')}
-      </IgnoreButton>
-      <DropdownMenu
-        size="sm"
-        className={className}
-        trigger={triggerProps => (
-          <DropdownTrigger
-            {...triggerProps}
-            aria-label={t('Ignore options')}
-            size={size}
-            icon={<IconChevron direction="down" />}
-            disabled={disabled}
-          />
-        )}
-        menuTitle={t('Ignore')}
-        items={dropdownItems}
-        isDisabled={disabled}
-      />
-    </ButtonBar>
-  );
-}
-
-export default IgnoreActions;
-
-const IgnoreButton = styled(Button)`
-  box-shadow: none;
-  border-radius: ${p => p.theme.borderRadius} 0 0 ${p => p.theme.borderRadius};
-`;
-
-const DropdownTrigger = styled(Button)`
-  box-shadow: none;
-  border-radius: 0 ${p => p.theme.borderRadius} ${p => p.theme.borderRadius} 0;
-  border-left: none;
-`;
